requirements.txt
h5py==2.10.0
imgaug==0.4.0
matplotlib==3.1.1
numpy==1.17.2
opencv-python==4.1.1.26
pandas==0.25.1
Pillow==6.1.0
scikit-image==0.15.0
scikit-learn==0.21.3
tqdm==4.48.2
tensorflow==2.3.0
efficientnet==1.1.1